Implement API for LanguageManager
